Docs and cleanup
